+++ test/fixtures/src/_data/snippet.js
@@ -1,0 +1,1 @@
+export default "Expanded snippet value";